feat: 多项功能改进和bug修复
- 修复抖音弹幕重连问题 (Cookie缓存导致)
- 优化HttpClient复用，避免socket耗尽
- 弹幕批量处理优化，减少UI线程压力
- 新增LiveStatusType枚举，支持直播/回放/离线状态
- 修复虎牙直播状态显示问题
- 关注页面显示回放状态(橙色标签)
- 恢复抖音搜索a_bogus签名支持

+++ AllLive.UWP/ViewModels/FavoriteVM.cs
@@ -105,8 +105,8 @@ namespace AllLive.UWP.ViewModels
                 {
                     LoaddingLiveStatus = false;
                     loadedCount = 0;
-                    // 排序，直播的在前面
-                    Items = new ObservableCollection<FavoriteItem>(Items.OrderByDescending(x => x.LiveStatus));
+                    // 排序：直播 > 回放 > 未直播
+                    Items = new ObservableCollection<FavoriteItem>(Items.OrderByDescending(x => (int)x.LiveStatus));
                 }
             }
         }
